+++ hc/front/tests/test_switch_channel.py
@@ -48,9 +48,6 @@ class SwitchChannelTestCase(BaseTestCase):
         self.assertEqual(r.status_code, 400)
 
     def test_it_allows_cross_team_access(self):
-        self.bobs_profile.current_project = None
-        self.bobs_profile.save()
-
         self.client.login(username="bob@example.org", password="password")
         r = self.client.post(self.url, {"state": "on"})
         self.assertEqual(r.status_code, 200)