+++ hc/front/tests/test_update_timeout.py
@@ -175,9 +175,6 @@ class UpdateTimeoutTestCase(BaseTestCase):
         self.assertEqual(r.status_code, 405)
 
     def test_it_allows_cross_team_access(self):
-        self.bobs_profile.current_project = None
-        self.bobs_profile.save()
-
         payload = {"kind": "simple", "timeout": 3600, "grace": 60}
 
         self.client.login(username="bob@example.org", password="password")